+++ src/openvpn/forward.c
@@ -42,6 +42,7 @@
 #include "dhcp.h"
 #include "common.h"
 #include "ssl_verify.h"
+#include "dco.h"
 
 #include "memdbg.h"
 
@@ -51,7 +52,6 @@ counter_type link_read_bytes_global;  /* GLOBAL */
 counter_type link_write_bytes_global; /* GLOBAL */
 
 /* show event wait debugging info */
-
 #ifdef ENABLE_DEBUG
 
 const char *
@@ -127,6 +127,25 @@ context_reschedule_sec(struct context *c, int sec)
         c->c2.timeval.tv_usec = 0;
     }
 }
+#if defined(ENABLE_LINUXDCO)
+static void
+check_dco_key_status(struct context *c)
+{
+
+
+    /* DCO context is not yet initialised or enab led */
+    if (!c->c2.did_open_tun || !dco_enabled(&c->options))
+    {
+        return;
+    }
+
+    struct tls_multi *multi = c->c2.tls_multi;
+    dco_context_t *dco = &c->c1.tuntap->dco_ctx;
+
+    dco_do_key_dance(dco, multi, c->options.ciphername);
+}
+#endif
+
 
 /*
  * In TLS mode, let TLS level respond to any control-channel
@@ -169,12 +188,20 @@ check_tls(struct context *c)
     }
 
     interval_schedule_wakeup(&c->c2.tmp_int, &wakeup);
+#if ENABLE_LINUXDCO
+    /* Our current code has no good hooks in the TLS machinery to install
+     * the keys to DCO. So we check/install keys after the whole TLS
+     * machinery has been completed
+     */
+    check_dco_key_status(c);
+#endif
 
     if (wakeup)
     {
         context_reschedule_sec(c, wakeup);
     }
 }
+
 
 void
 check_tls_errors_co(struct context *c)
@@ -1093,6 +1120,15 @@ process_incoming_link(struct context *c)
     perf_pop();
 }
 
+#if ENABLE_LINUXDCO
+static void
+process_incoming_dco(struct context *c)
+{
+    msg(M_INFO, __func__);
+    ovpn_do_read_dco(&c->c1.tuntap->dco_ctx);
+}
+#endif
+
 /*
  * Output: c->c2.buf
  */
@@ -1614,9 +1650,17 @@ process_outgoing_link(struct context *c)
                 socks_preprocess_outgoing_link(c, &to_addr, &size_delta);
 
                 /* Send packet */
-                size = link_socket_write(c->c2.link_socket,
-                                         &c->c2.to_link,
-                                         to_addr);
+#if ENABLE_LINUXDCO
+                if (c->c2.link_socket->info.dco_installed)
+                {
+                    size = ovpn_do_write_dco(&c->c1.tuntap->dco_ctx, c->c2.tls_multi->peer_id, &c->c2.to_link);
+                }
+                else
+#endif
+                {
+                    size = link_socket_write(c->c2.link_socket, &c->c2.to_link,
+                                             to_addr);
+                }
 
                 /* Undo effect of prepend */
                 link_socket_write_post_size_adjust(&size, size_delta, &c->c2.to_link);
@@ -1878,6 +1922,9 @@ io_wait_dowork(struct context *c, const unsigned int flags)
 #ifdef ENABLE_ASYNC_PUSH
     static int file_shift = 8;     /* listening inotify events */
 #endif
+#ifdef ENABLE_LINUXDCO
+    static int dco_shift = 10;    /* Event from DCO module */
+#endif
 
     /*
      * Decide what kind of events we want to wait for.
@@ -1985,6 +2032,12 @@ io_wait_dowork(struct context *c, const unsigned int flags)
      */
     socket_set(c->c2.link_socket, c->c2.event_set, socket, (void *)&socket_shift, NULL);
     tun_set(c->c1.tuntap, c->c2.event_set, tuntap, (void *)&tun_shift, NULL);
+#if ENABLE_LINUXDCO
+    if (socket & EVENT_READ && c->c2.did_open_tun)
+    {
+        dco_event_set(&c->c1.tuntap->dco_ctx, c->c2.event_set, (void *) &dco_shift);
+    }
+#endif
 
 #ifdef ENABLE_MANAGEMENT
     if (management)
@@ -2107,4 +2160,13 @@ process_io(struct context *c)
             process_incoming_tun(c);
         }
     }
+#if ENABLE_LINUXDCO
+    else if (status & DCO_READ)
+    {
+        if(!IS_SIG(c))
+        {
+            process_incoming_dco(c);
+        }
+    }
+#endif
 }