+++ src/openvpn/ssl_common.h
@@ -145,6 +145,12 @@ enum ks_auth_state {
                                 */
 };
 
+enum dco_key_status {
+    DCO_NOT_INSTALLED,
+    DCO_INSTALLED_PRIMARY,
+    DCO_INSTALLED_SECONDARY
+};
+
 /**
  * Security parameter state of one TLS and data channel %key session.
  * @ingroup control_processor
@@ -216,6 +222,8 @@ struct key_state
     time_t acf_last_mod;
     char *auth_control_file;
     char *auth_pending_file;
+
+    enum dco_key_status dco_status;
 };
 
 /** Control channel wrapping (--tls-auth/--tls-crypt) context */
@@ -606,6 +614,10 @@ struct tls_multi
     /**< Array of \c tls_session objects
      *   representing control channel
      *   sessions with the remote peer. */
+
+    /* Only used when DCO is used to remember how many keys we installed
+     * for this session */
+    int dco_keys_installed;
 };
 
 /**  gets an item  of \c key_state objects in the