+++ Changes.rst
@@ -39,6 +39,13 @@ AEAD implicit IV using full IV size
     packet-id is scrambled and the number of packet already sent per key cannot
     be inferred from a single packet.
 
+Data channel offloading with ovpn-dco
+    2.6.0+ implements support for data-channel offloading where the data packets
+    are directly processed and forwarded in kernel space thanks to the ovpn-dco
+    kernel module. The userspace openvpn program acts purely as a control plane
+    application.
+    
+
 Deprecated features
 -------------------
 ``inetd`` has been removed